feat: move remaining atom to components lib
