+++ packages/components/src/atoms/FSelectInput/FSelectInput.constants.ts
@@ -1,0 +1,12 @@
+export const TRANSACTION_TYPES = [
+  "Depósito",
+  "Saque",
+  "Transferência",
+  "Pagamento",
+  "Investimento",
+  "Empréstimo",
+  "Financiamento",
+  "Câmbio de moeda",
+  "DOC/TED",
+  "PIX",
+];
