+++ packages/components/src/index.ts
@@ -5,3 +5,4 @@
 export { FButton } from "./atoms/FButton/FButton.tsx";
 export { FIconButton } from "./atoms/FIconButton/FIconButton.tsx";
 export { FInput } from "./atoms/FInput/FInput.tsx";
+export { FSelectInput } from "./atoms/FSelectInput/FSelectInput.tsx";
