--- packages/dashboard/src/components/atoms/FSelectInput/FSelectInput.stories.tsx
@@ -1,23 +1,0 @@
-import { Box } from "@mui/material";
-import type { Meta, StoryFn } from "@storybook/react";
-import { fn } from "@storybook/test";
-import SelectInput from "./FSelectInput";
-
-const meta: Meta<typeof SelectInput> = {
-  title: "Atoms/SelectInput",
-  component: SelectInput,
-  parameters: {
-    layout: "centered",
-  },
-  args: { onChange: fn() },
-};
-
-export default meta;
-
-export const Default: StoryFn = (args) => {
-  return (
-    <Box width={500}>
-      <SelectInput {...args} />
-    </Box>
-  );
-};
