+++ packages/dashboard/src/components/molecules/FTransactionForm/FTransactionForm.tsx
@@ -1,5 +1,4 @@
 "use client";
-import FSelectInput from "@/components/atoms/FSelectInput/FSelectInput";
 import {
   Transaction,
   TransactionData,
@@ -7,7 +6,7 @@ import {
 } from "@/services/Transaction/Transaction.model";
 import { Box, SelectChangeEvent } from "@mui/material";
 import Stack from "@mui/material/Stack";
-import { FButton, FInput } from "components";
+import { FButton, FInput, FSelectInput } from "components";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 export interface FTransactionFormProps {
