+++ packages/dashboard/src/constants.ts
@@ -29,16 +29,3 @@ export const MENU_ITEMS_HOME: MenuItemType[] = [
     path: "/servicos",
   },
 ];
-
-export const TRANSACTION_TYPES = [
-  "Depósito",
-  "Saque",
-  "Transferência",
-  "Pagamento",
-  "Investimento",
-  "Empréstimo",
-  "Financiamento",
-  "Câmbio de moeda",
-  "DOC/TED",
-  "PIX",
-];
